+++ TSS.Java/src/tss/tpm/TPMT_SYM_DEF_OBJECT.java
@@ -35,7 +35,7 @@ public class TPMT_SYM_DEF_OBJECT extends TpmStructure
     public TPMT_SYM_DEF_OBJECT(TPM_ALG_ID _algorithm, int _keyBits, TPM_ALG_ID _mode)
     {
         algorithm = _algorithm;
-        keyBits = (short)_keyBits;
+        keyBits = _keyBits;
         mode = _mode;
     }
